app/backend/requirements.txt
# Runtime
openai
pydantic
numpy
pandas
scikit-learn
sentence-transformers
kagglehub
python-dotenv

# API server
fastapi
uvicorn[standard]

# Testing
httpx
pytest
pytest-asyncio
anyio[trio]
